valaccodebasemodule.vala: Allow unref_function to be a macro
Some APIs define unref_function to be a macro of the form
do { ... } while (0).

valac uses the unref_function as an expression, but a do loop is a
statement. This causes gcc to exit with a syntax error.

To fix this, destroy_value should return a CCodeStatement or a
CCodeFragment containing CCodeStatements, rather than a CCodeExpression.

https://bugzilla.gnome.org/show_bug.cgi?id=785570